Refactor, separate out ApiController into smaller classes
